Fixed timesetting for hwclock localtime users
When we set system time from the BIOS clock via hwclock we must honor
the timezone the user previous selected during interactive_time().
After boot we are in an UTC environment. If we now  - for localtime
users - parse hwclcock value to systemtime, then we get later a false
systemtime when copying the TZ file to /etc/localtime. Changing
/etc/localtime always recalculate the time according to the timezone.

Signed-off-by: Gerhard Brauer <gerbra@archlinux.de>