Modified timesetting procedure again
With our current code we get problems when the user want to set datetime
manually in our dialog. Revisiting last changes i see: it's a bad thing
to force any timesetting based on a *selected* timezone in an still UTC
environment.
So the easiest way (and i hope the last change on this!) is:
a) user selected a timezone
b) we set this timezonefile to /etc/localtime
c) Now we ask: is your hwclock on UTC or localtime?
c) UTC BIOSclock user now have both correct system and hw time.
d) For localtime users we must re-read the hwclock into our new
timezone, cause copy'ing above timezone file has recalculated the
systemtime to a false time.
I've tested these commit on my PC, both with hwclock on UTC and
localtime, also with manual setting and via ntp(date) in the aif dialog.

Signed-off-by: Gerhard Brauer <gerbra@archlinux.de>