make AIF track needed filesystem packages
filesystem packages are scheduled to go out of base, hence
AIF must track which ones are needed and which are not by keeping an eye
open during the filesystems creation (and rollback) steps.
In the interactive procedure, the appropriate packages will be
preselected, in the automatic procedure, you're supposed to configure
the needed packages yourself now.  This is reflected in the examples.