btrfs: ask for mountpoint
This commit supports using a btrfs directly.
Despite recommendations and strong suggestions to somehow prevent users from
doing this (and using subvolumes instead) - see
http://mailman.archlinux.org/pipermail/arch-releng/2010-December/001341.html -
it's taking way to long for code to support this to go into AIF.
It's now the users' choice to choose what they want and deal with the
consequences.